workers/pyproject.toml
[project]
name = "dreambig-workers"
version = "0.1.0"
description = "Game ranking scraper workers for DreamBig"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "psycopg[binary]>=3.1",
    "redis>=5.0",
    "rq>=1.16",
    "apscheduler>=3.10",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "jieba>=0.42",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
